Funcionalidad a Entradas

+++ src/Modelo/Sistema_inventario.sql
@@ -93,7 +93,7 @@ CREATE TABLE Entradas (
 	ID INT UNSIGNED AUTO_INCREMENT NOT NULL,
     ID_proveedor INT UNSIGNED NOT NULL,
     Fecha DATE NOT NULL, 
-    Cantidad_producto INT NOT NULL,
+    Cantidad_producto DOUBLE NOT NULL,
     
     PRIMARY KEY (ID),
     foreign key (ID_proveedor) references Proveedores (ID)
